+++ app/lib/main.dart
@@ -4,10 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'blocks/form.dart';
-
-// providerのgolabal宣言
-final formProvider = Provider((_) => 0);
+import 'components/pages/form_page.dart';
 
 void main() {
   runApp(
@@ -26,22 +23,7 @@ class MyApp extends ConsumerWidget {
       theme: ThemeData(
         primarySwatch: Colors.green,
       ),
-      home: MyHomePage(),
+      home: FormPage(),
     );
   }
 }
-
-class MyHomePage extends ConsumerWidget {
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Tetris評価サーバ"),
-      ),
-      body: Center(
-        child: Container(margin: const EdgeInsets.all(40), child: SubmitForm()),
-      ),
-    );
-  }
-}
-
