+++ app/lib/model/form_model.dart
@@ -1,0 +1,7 @@
+class FormModel {
+  final String repository_url;
+  final String branch_name;
+  final String level;
+
+  FormModel(this.repository_url, this.branch_name, this.level);
+}
